requirements.txt
optuna>=3.3.0
optuna-dashboard[preferential]>=0.13.0
pillow
optuna-fast-fanova
gunicorn
requests

